쿼리 추가 및 회원가입 뷰단 유효성검사 이름까지 완료

+++ src/test/java/footBall/user/UserServiceImplTest.java
@@ -19,7 +19,7 @@ class UserServiceImplTest {
         UserRequest dto = new UserRequest();
         dto.setFbUserEmail("chltmdals@naver.com");
         dto.setFbUserPswd("12345");
-        Long id = userService.login(dto);
+        int id = userService.login(dto);
         System.out.println(id);
     }
 }
